+++ packages/adapter-netlify/index.js
@@ -24,6 +24,7 @@ export default function () {
 				entryPoints: ['.svelte-kit/netlify/entry.js'],
 				outfile: join(functions, 'render/index.js'),
 				bundle: true,
+				inject: [join(files, 'shims.js')],
 				platform: 'node'
 			});
 
@@ -39,6 +40,7 @@ export default function () {
 			utils.copy_client_files(publish);
 
 			utils.log.minor('Writing redirects...');
+			utils.copy('_redirects', `${publish}/_redirects`);
 			appendFileSync(`${publish}/_redirects`, '\n\n/* /.netlify/functions/render 200');
 		}
 	};