+++ packages/adapter-node/index.d.ts
@@ -1,3 +1,10 @@
-declare function plugin(options?: { out?: string }): import('@sveltejs/kit').Adapter;
+declare function plugin(options?: {
+	out?: string;
+	precompress?: boolean;
+	env?: {
+		host?: string;
+		port?: string;
+	};
+}): import('@sveltejs/kit').Adapter;
 
 export = plugin;