+++ packages/kit/rollup.config.js
@@ -9,7 +9,9 @@ import pkg from './package.json';
 const external = [].concat(
 	Object.keys(pkg.dependencies || {}),
 	Object.keys(pkg.peerDependencies || {}),
-	Object.keys(process.binding('natives'))
+	Object.keys(process.binding('natives')),
+	'typescript',
+	'svelte2tsx'
 );
 
 export default [