+++ packages/kit/test/apps/options/svelte.config.js
@@ -1,5 +1,6 @@
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
+	extensions: ['.jesuslivesineveryone', '.whokilledthemuffinman', '.svelte.md', '.svelte'],
 	kit: {
 		files: {
 			assets: 'public',